# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API Engine driving the StreamingPlayer media pipeline.",
        "title": "StreamingPlayer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/media": {
            "get": {
                "description": "Retrieve a paginated list of cataloged video items.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "List all library media",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of media",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/media/{id}": {
            "get": {
                "description": "Find metadata of a catalog item by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Retrieve single media item details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Media not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}